Use exception guards when fetching Windows info
Calling the [] method of VendorDictionary either returns a non-nullable type if
the key exists or throws an exception if the key does not exist.
This change fixes the Windows platform code and makes sure that if a dictionary
key does not exist, null is returned instead of the library throwing an error,
matching the behaviour of the UNIX platform code.